Add ehcache-jcache project
This scratch project demonstrates the use of the JSR-107 compliant
version of ehcache with Spring's caching abstraction.

The test infrastructure has been replicated to use a different ehcache
version.

Problems encountered:
* "default" is a reserved keyword in the ehcache world. Registering
a cache with that name does not seem to update the bridge properly
* By default, the cache accepts only Serializable value as it copies
the value instead of keeping the reference. This has been disabled