This commit provides a spring boot starter for the HornetQ JMS broker.
HornetQAutoConfiguration provides a javax.jms.ConnectionFactory that
integrates with HornetQ if none is present and if the necessary HornetQ
classes are present in the classpath.

The connection factory connects to a broker available on the local
machine by default. A configuration switch allows to enable an embedded
mode that starts HornetQ as part of the application.

In such a mode, the spring.hornetq.embedded.* properties provide
additional options to configure the embedded broker.