Add basic auto-configuration
Add the most basic auto-configuration we could find. It creates a
ConnectionFactory on the local broker if none exists and if HornetQ is
available.

We also make sure to run this auto-config *before* the general JMS
auto-configuration so that the JmsTemplate can be created properly.

We also add a test to make sure that everything is working as expected.

TODO: rename sample-app to hornetq-sample-app and fix the name